renamed unsavedChanges
